src/redis/val_string_test.go
package redis

import (
	"testing"
	"time"

	"store"
	"testing_helpers"
)

// tests the string value
func TestStringValue(t *testing.T) {
	s := setupRedis()
	src := NewString("blake", time.Now())

	b, err := s.SerializeValue(src)
	if err != nil {
		t.Fatalf("Unexpected serialization error: %v", err)
	}

	val, vtype, err := s.DeserializeValue(b)
	if err != nil {
		t.Fatalf("Unexpected deserialization error: %v", err)
	}
	if vtype != STRING_VALUE {
		t.Fatalf("Unexpected value type enum: %v", vtype)
	}
	dst, ok := val.(*String)
	if !ok {
		t.Fatalf("Unexpected value type: %T", val)
	}

	testing_helpers.AssertEqual(t, "data", src.data, dst.data)
	testing_helpers.AssertEqual(t, "time", src.time, dst.time)
}

// tests that the tombstone struct satisfies the
// value interface
func TestStringInterface(t *testing.T) {
	func (store.Value){}(NewTombstone(time.Now()))
}
